Moved API call to App.jsx so that distributing data and states is easier
