+++ src/components/MovieList.jsx
@@ -4,42 +4,11 @@ import MovieCard from "./MovieCard";
 import '../styles/MovieList.css'
 import '../styles/LoadButton.css'
 
-//API Info for fetch
-const API_KEY = import.meta.env.VITE_API_KEY
-const options = {method: 'GET', headers: {accept: 'application/json',
-                Authorization: `Bearer ${API_KEY}`}}
-
-//Path info for movie posters
+////Path info for movie posters
 const imgURL = 'https://image.tmdb.org/t/p'
 const posterSize = '/w500'
 
-export default function MovieList() {
-
-    //Store state of movie data and current page num
-    const [movieData, setMovieData] = useState([])
-    const [pageNum, setPageNum] = useState(1)
-
-    //URL for featching data from API
-    const URL = `https://api.themoviedb.org/3/movie/now_playing?language=en-US&page=${pageNum}`
-
-    useEffect(() => {
-        const fetchMovieData =  async () => {
-            try{
-                var res = await fetch(URL, options)
-                if(res.ok){
-                    const data = await res.json();
-                    setMovieData([...movieData, ...data.results]);
-                }else{
-                    throw new Error("API Not Responding")
-                }
-            }catch(error){
-                console.error("Error: ", error.message)
-            }
-        }
-        fetchMovieData();
-    },[pageNum])
-
-    const load = () => setPageNum(pageNum + 1)
+export default function MovieList({data,load}) {
 
     //Modal state data
     const [show, setShow] = useState(false);
@@ -47,14 +16,14 @@ export default function MovieList() {
 
     const close = () => setShow(false);
     const open = id => event => {
-        setMov(movieData.find(movie => movie.id === id))
+        setMov(data.find(movie => movie.id === id))
         setShow(true);
 
     }
     return (
         <>
         <div className="MovieList">
-            {movieData.map((movie) => {
+            {data.map((movie) => {
                 return <MovieCard loadModal={open(movie.id)} title={movie.title} image={imgURL + posterSize + movie.poster_path} rating={movie.vote_average} key={movie.id}/>
             })};
         </div>
